Code review for User Stroy # 2433 "Go over ovs_rec_for_each and optimize for route table lookup (zebra add route ptr to rib structure)"
Problem Cause:-
When we update the status of the route's next-hop being programmed in the rib, we walk all the entries in the OVSDB route table. This is an expensive operation since the OSDB route DB contains all the next-hops for a route and all IPV4 and IPV6 next-hops.

Solution:-
To reduce the time spent in walking the route DB to get to a given route's next-hop, we cache the pointer to the corresponding route-row in the OVSDB route table at the time when we are creating the next-hop for the route.

Change-Id: I2d7b473789d43d6b731607bc54fed4beb4234f85
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hp.com>