Code diff for user story # 2793
The following user stories are being committed as part of this review:-
1. Garbage collect static routes/NH on port/vrf delete
2. Need to listen to interface UP/DOWN notification from OVSDB for removing /adding the next-hops for static routes.

We maitain a L3 port cache to detect if any L3 ports got added or deleted. If the L3 ports got deleted we remove the static route configuration from the OVSDB. In case the interface is shut/un-shut we mark the next-hop/route as unselected or selected.

Change-Id: Id61e835c8d39e7092f82d5714309d14afb89c65a
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>