fix: usr: Tagia Bug#609 Zebra adding a next hop when interface is down
Zebra backend does not call the OVSDB route and next-hop selected bit set
function in all cases when zebra is trying to modify the kernel with the
route and its next-hops. In the fix, we are trying to call the route and
next-hop update function in all scenarios.
The CT script also adds test cases as reported in the bug.

Change-Id: Ib274165c23d45709136fb9116f9c47deb9f8f94b
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>