fix: usr: Tagia Bug#610 ops-zebra crash while dereferencing null pointer
While reading the OVSDB route table ops-zebra sometimes crashes if the
next-hops for a given route are pointing to null. ops-zebra attempts to
dereferece the null pointer and crashes. The fix tests if the next-hop
array within a route is null or not. If the next-hop array is a null pointer
then ops-zebra does not process that route.

Change-Id: Ie0e39d7b81fa78803885435a3dae74add0bdd62f
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>