bgpd: Modernise BGP defaults for MRAI and connect time
Quagga, and ops-quagga as a result, uses historic BGP defaults for
its minimum route advertisement interval (MRAI) timers, and its
ConnectRetry timer.  It uses 30s and 5s for the eBGP and iBGP MRAIs,
and 120s for the ConnectRetry timer.  These values are quite high,
and delay convergence and session establishment, and are unlikely to
be desirable in modern networks, data-centres especially.

Additionally, Quagga has a start-timer that adds a delay to starting
a session of 5s.  This timer is further subject to psuedo-randomised
jitter of +/- (between 0 and the timer value.

The MRAI can, without a doubt, be significantly reduced.  The optimal
MRAI is related to the size of the network, its diameter in terms of
BGP propagation latency particularly.  Prior research suggests values
of 5s to 15s for Internet BGP, as reasonably conservative values.
Further, other implementations have long shipped with 1s or even 0s
MRAI values.

A too low MRAI, e.g.  a 0 MRAI, can greatly increase the number of
BGP messages a speaker sends, and hence the bandwidth and even CPU it
might use.  I.e.  some MRAI is better than no MRAI, with current BGP,
or it is prone to excessive - even extreme - chattiness.

This commit sets the defaults to:

- 3s for the eBGP MRAI
- 1s for iBGP MRAI (lower, but non-0, would be nice for this)
- 1s for the Start timer (so +/- 1s jitter).
- 5s for the ConnectRetry timer

Change-Id: Iffe05540dec7b12f4a2dd3a5a5beabe9a79721fd
Signed-off-by: Paul Jakma <paul.jakma@hpe.com>