ops-zebra crash while dereferencing null pointer
While reading the OVSDB route table ops-zebra sometimes crashes if the
next-hops for a given route are pointing to null. ops-zebra attempts to
dereferece the null pointer and crashes. The fix tests if the next-hop
array within a route is null or not. If the next-hop array is a null pointer
then ops-zebra does not process that route.

TG-1016
Tags: fix, usr

Change-Id: I939917c1c994f55a281ed79f1678511b4afc4c06
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>