Handle request for immediate run of OVSDB loop.
Taiga # 1150: Zebra stops processing Route (and other OVSDB tables)

The handling for the poll_loop timer was incorrect in the case where
an immediate re-run of the OVSDB loop is required. In this case, the
timeout_when value is LLONG_MIN (very negative). Handle this case
explicitly by looping instead of starting a timer.

Change-Id: I3715a0b0850d5130b46977ce7d82c15d8a4d06a7
Signed-off-by: DanCripe <dan@quattronetworks.com>