ops-zebra not cleaning up OVSDB in case of "no routing"
THis fix addresses a corner case of "no routing" when the L3 interface
does not have an IP/IPv6 address but this interface is L3 since it is the
part of vrf. In case of "no routing" in this kind of interface, zebra should
clean-up the DB for the route going via this interface.

Change-Id: I63ec978863547f1c7c6953e74788a8707ef34b53
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>