ops-zebra not cleaning up OVSDB in case of "no routing"
THis fix addresses a corner case of "no routing" when the L3 interface
does not have an IP/IPv6 address but this interface is L3 since it is the
part of vrf. In case of "no routing" in this kind of interface, zebra should
clean-up the DB for the route going via this interface.

TG-1502
Tags: fix, usr

Change-Id: I421819be7a85f10b38d497da7cadf6a76e8373ac
Signed-off-by: Gaurav Gupta <gaurav.gupta4@hpe.com>