Loopback interface creation in ospf
During loopback interface configuration in ospf,
there could be a situation when the kernel
loopback interface is not ready/created yet by
ops-portd. In such a case, the interface hangs
in ops-ospfd with ifindex equal to 0. A special
handling added for that case: create loopback
interface during update event for the port entry.
The port entry (kernel_interface_ready field set
to true) is updated by ops-portd when loopback
interface is created in kernel.

Tags: dev, fix
TG-1986

Change-Id: I971c214f97a59e139c02d1212e78c834d964f02e
Signed-off-by: Nikolay Stroykov <stroykov@mera.ru>