Default route LSA refresh
Fix LSA refresh for redistributed LSAs. After default route
configuration and setup as distributed, it is distributed as expected.
But after LSA refresh timer expires it is not refreshed and the
corresponding LSA is remove from LSADB. This problem relates to all
LSAs distributed as AS-External.

Tags: fix, dev, usr

Change-Id: I82f573762837f217986dd45212045144e05e3d30
Signed-off-by: Nikolay Stroykov <stroykov@mera.ru>